remove(payment): usecase interface 삭제
